+++ src/Simu/Loi.java
@@ -4,7 +4,7 @@ import java.util.Random;
 
 public class Loi {
 
-    public double getLoiExponentionnelleTelephone(int time){
+    public double getLoiExponentionnelleTelephone(double time){
         double res=0;
         double lambda=0;
         //loi exponentielle espÃ©rance 5
@@ -22,7 +22,7 @@ public class Loi {
         res= -((1/ lambda) * Math.log( 1- var));
         return res;
     }
-    public double getLoiExponentielleMail(int time){
+    public double getLoiExponentielleMail(double time){
         double res=0;
         double lambda=0;
         //loi exponentielle espÃ©rance 5
@@ -50,6 +50,7 @@ public class Loi {
     public int getDureeTraitementAppel(){
         return loiNormale(300,900);
     }
+
     public int loiNormale(int min, int max){
         int res=0;
         Random rand = new Random();
